+++ R/InfiniumDiffMetMotR.R
@@ -1,0 +1,87 @@
+#' InfiniumDiffMetMotR.
+#' 
+#' @name InfiniumDiffMetMotR
+#' @docType package
+NULL
+
+#' Annotation of InfiniumMethylation450 beadsChip
+#' 
+#' a dataset containing the annotation of each probe. The variables are as follows:
+#' 
+#' \itemize{
+#' \item IlmnID. probe ID
+#' \item Genome_Build. version of genome
+#' \item CHR. chromosome
+#' \item MAPINFO. coordinate of genome
+#' UCSC_RefGene_Group. Annotation against gene
+#' Relation_to_UCSC_CpG_Island. Annotation against CpG Island
+#' }
+#' 
+#' @docType data
+#' @keywords datasets 450 Methyl450 Infinium annotation
+#' @name Methyl450anno
+#' @usage data(Methyl450anno)
+#' @format A data frame with 486428 rows and 6 variables
+NULL
+
+#' Annotation of InfiniumMethylationEPIC beadsChip
+#' 
+#' a list containing the annotation of each probe. The variables are as follows:
+#' 
+#' \itemize{
+#' \item IlmnID. probe ID
+#' \item Genome_Build. version of genome
+#' \item CHR. chromosome
+#' \item MAPINFO. coordinate of genome
+#' UCSC_RefGene_Group. Annotation against gene
+#' Relation_to_UCSC_CpG_Island. Annotation against CpG Island
+#' }
+#' 
+#' @docType data
+#' @keywords datasets EPIC 850 Infinium annotation
+#' @name EPICanno
+#' @usage data(EPICanno)
+#' @format A data frame with 867531 rows and 6 variables
+NULL
+
+#' IMAGE PWM database
+#' 
+#' a list containing the annotation of each probe. 
+#' 
+#' @docType data
+#' @keywords datasets IMAGE PWM list
+#' @name IMAGE_PWMlist
+#' @usage data(IMAGE_PWMlist)
+#' @format A list with 1611 elements
+NULL
+
+#' Annotation of InfiniumMethylationEPIC beadsChip
+#' 
+#' a list containing the annotation of each probe. The variables are as follows:
+#' 
+#' \itemize{
+#' \item IlmnID. probe ID
+#' \item Genome_Build. version of genome
+#' \item CHR. chromosome
+#' \item MAPINFO. coordinate of genome
+#' UCSC_RefGene_Group. Annotation against gene
+#' Relation_to_UCSC_CpG_Island. Annotation against CpG Island
+#' }
+#' 
+#' @docType data
+#' @keywords datasets EPIC 850 Infinium annotation
+#' @name EPICanno
+#' @usage data(EPICanno)
+#' @format A data frame with 867531 rows and 6 variables
+NULL
+
+#' EPIC control database
+#' 
+#' EPIC control dataset from wateRmelon 
+#' 
+#' @docType data
+#' @keywords wateRmelon EPIC
+#' @name epic.controls
+#' @usage data(epic.controls)
+#' @format 635 abs. of 4 variables
+NULL