+++ R/seqExtract.R
@@ -1,4 +1,20 @@
-seqExtract<- function(positions, genome , seq_range = c(-5000,5000)){ #sequence extraction
+#' Sequence extraction
+#' 
+#' Sequence extraction from the given range from position.
+#' 
+#' @param positions genpomic coordinates
+#' @param genome genome sequence (from BSgenome data)
+#' @param seq_range sequence extraction target range from position
+#' 
+#' @importFrom Biostrings DNAStringSet getSeq
+#' @importFrom GenomeInfoDb seqlengths
+#' 
+#' @return a DNAStringSet format sequences
+#' 
+#' @keywords genome sequence
+#' @export
+
+seqExtract<- function(positions, genome, seq_range = c(-5000, 5000)){ #sequence extraction
   seq_length <- seqlengths(genome)
   positions <- positions[(positions[,1] != ""), ,drop=F]
   chrom <- positions[,1]