+++ CHANGELOG.md
@@ -17,3 +17,14 @@
 ### Fixed
 - Numer of core to be used has been changed from 16 to 4.
 
+## [1.5.0] - 2019-10-05
+### Update
+- Format of scripts was hanged to Roxygen format.
+
+### Added
+- When using the sampling option, indicate it on standerd out
+- Output DMPs
+- idat files support as the input files
+
+### Fixed
+- Error with the case of DMR = 0/1